reorganized views into view folder
