+++ src/components/views/MainView.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Button } from "react-bootstrap";
-import { DegreePlan } from "../interfaces/degreeplan";
-import { DegreePlanList } from "./DegreePlanList";
-import { CSVExport } from "./CSVExport";
-import invalidSemester from "../exampleData/invalid_semester.json";
-import { Semester } from "../interfaces/semester";
+import { DegreePlan } from "../../interfaces/degreeplan";
+import { DegreePlanList } from "../DegreePlanList";
+import { CSVExport } from "../CSVExport";
+import invalidSemester from "../../exampleData/invalid_semester.json";
+import { Semester } from "../../interfaces/semester";
 
 // button for switching to the PlanView
 function PlanViewButton({
