+++ src/components/views/PlanView.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import { Button } from "react-bootstrap";
-import { DegreePlan } from "../interfaces/degreeplan";
-import { Semester } from "../interfaces/semester";
-import { AddSemesterForm } from "./AddSemesterForm";
-import { SemesterScrollBox } from "./SemesterScrollBox";
-import invalidSemester from "../exampleData/invalid_semester.json";
+import { DegreePlan } from "../../interfaces/degreeplan";
+import { Semester } from "../../interfaces/semester";
+import { AddSemesterForm } from "../AddSemesterForm";
+import { SemesterScrollBox } from "../SemesterScrollBox";
+import invalidSemester from "../../exampleData/invalid_semester.json";
 import "./components.css";
 
 // Button for switching to the SemesterView after selecting a semester
